+++ src/posts/posts.service.ts
@@ -3,10 +3,14 @@ import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
 import { PostsRepository } from './posts.repository';
 import { NotFoundError } from '../errors/notfoundError';
+import { PublicationsRepository } from '../publications/publications.repository';
+import { ForbiddenError } from '../errors/forbiddenError';
 
 @Injectable()
 export class PostsService {
-  constructor(private postRepository: PostsRepository) { }
+  constructor(
+    private postRepository: PostsRepository,
+    private publicationRepository: PublicationsRepository) { }
 
   async createPost(createPostDto: CreatePostDto) {
     return await this.postRepository.createPost(createPostDto);
@@ -31,6 +35,10 @@ export class PostsService {
   async removePost(id: number) {
     const post = await this.postRepository.findOnePost(id);
     if (!post) throw new NotFoundError(id);
+
+    const publication = await this.publicationRepository.findOnePublicationByPostId(id);
+    if (publication) throw new ForbiddenError(id);
+
     return await this.postRepository.removePost(id);
   }
 }
